Remove `notification` column from su.sqlite import.
* It seems that SuperUser somehow recognizes that our fresh DB is new,
  likely because it somehow marks the DB when the app itself creates it.
On first startup, it assumes our fresh DB is old, and tried to alter it
and add the notifications column, which causes the Settings app to
crash. We can avoid this crash by removing the `notification` column
ourselves, and allowing SuperUser to add it back the first time it's
launched.